fix: remove .mobile-press stylesheet and show message actions on hover regardless of the screen size (#138)
Historically, mobile-press class was appended programmatically whenever a user tapped on a message in the message list. The goal of it was to make message actions visible.
After an internal discussion, we concluded that this isn't optimal and introduces additional complexity. We decided to remove this behavior and rely on CSS `:focus-within` pseudo-selector which is already in place within the SDKs and handles the desired goal quite well.

This change could be potentially breaking for some customers who might have written their own stylesheets/selectors based on the existence of `.mobile-press`. Our recommendation is to rewrite those stylesheets/selectors with `:focus` or `:focus-within` pseudo-selectors on mind.