fix: make message actions keyboard accessible (#272)
### 🎯 Goal

Currently message actions are only displayed on hover, and are not
accessible from keyboard. This PR makes message actions more accessible.

### 🛠 Implementation details

1. Message actions are now displayed when the message is focused
2. Action buttons now have outline to clearly show focus

This PR includes a slightly scary change to the `button-reset` mixin.
Previously it removed outline. However, having outline for focused items
is important for accessibility, so now outline is not removed.

### 🎨 UI Changes


https://github.com/GetStream/stream-chat-css/assets/975978/7824625d-5ba0-4ba1-b4b5-1572194e33ac