docs: simplify CSS docs (#309)
### 🎯 Goal

With the new documentation design, we want to have nice and crisp docs.
The docs of CSS variables have complex tables that are hard to
understand, and harder to style. To help the docs effort, we're creating
a simplified version of the docs.

If customers don't like the simplified version, we can restore it.